refactor: ホーム画面の利用規約・プライバシーポリシーをHomeFooterのコンポーネントに変更
